Allow much larger ids in postgres db

+++ resources/migrations/20150605-01-create-sources.up.sql
@@ -1,10 +1,10 @@
-CREATE TABLE sources (id INTEGER NOT NULL,
-                      results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE sources (id BIGINT NOT NULL,
+                      results_id BIGINT REFERENCES results (id) NOT NULL,
                       PRIMARY KEY (results_id, id),
                       name TEXT,
                       vip_id TEXT,
                       datetime DATE,
                       description TEXT,
                       organization_url TEXT,
-                      feed_contact_id INT,
+                      feed_contact_id BIGINT,
                       tou_url TEXT);
