+++ resources/migrations/20150605-02-create-states.up.sql
@@ -1,5 +1,5 @@
-CREATE TABLE states(id INTEGER NOT NULL,
-                    results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE states(id BIGINT NOT NULL,
+                    results_id BIGINT REFERENCES results (id) NOT NULL,
                     PRIMARY KEY (results_id, id),
                     name TEXT,
-                    election_administration_id INT);
+                    election_administration_id BIGINT);
