+++ resources/migrations/20150605-04-create-election-officials.up.sql
@@ -1,5 +1,5 @@
-CREATE TABLE election_officials (id INTEGER NOT NULL,
-                                 results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE election_officials (id BIGINT NOT NULL,
+                                 results_id BIGINT REFERENCES results (id) NOT NULL,
                                  PRIMARY KEY (results_id, id),
                                  name TEXT,
                                  title TEXT,
