+++ resources/migrations/20150605-09-create-precinct-split-electoral-districts.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE precinct_split_electoral_districts(results_id INTEGER REFERENCES results (id) NOT NULL,
-                                                precinct_split_id INTEGER,
-                                                electoral_district_id INTEGER);
+CREATE TABLE precinct_split_electoral_districts(results_id BIGINT REFERENCES results (id) NOT NULL,
+                                                precinct_split_id BIGINT,
+                                                electoral_district_id BIGINT);
