+++ resources/migrations/20150605-13-create-electoral-districts.up.sql
@@ -1,5 +1,5 @@
-CREATE TABLE electoral_districts (id INTEGER NOT NULL,
-                                  results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE electoral_districts (id BIGINT NOT NULL,
+                                  results_id BIGINT REFERENCES results (id) NOT NULL,
                                   PRIMARY KEY (results_id, id),
                                   name TEXT,
                                   type TEXT,
