+++ resources/migrations/20150605-16-create-ballot-responses.up.sql
@@ -1,5 +1,5 @@
-CREATE TABLE ballot_responses (id INTEGER NOT NULL,
-                               results_id INTEGER REFERENCES results (id) NOT NULL,
+CREATE TABLE ballot_responses (id BIGINT NOT NULL,
+                               results_id BIGINT REFERENCES results (id) NOT NULL,
                                PRIMARY KEY (results_id, id),
                                text TEXT,
                                sort_order INTEGER);
