+++ resources/migrations/20150605-18-create-referendum-ballot-responses.up.sql
@@ -1,4 +1,4 @@
-CREATE TABLE referendum_ballot_responses (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                          referendum_id INTEGER,
-                                          ballot_response_id INTEGER,
+CREATE TABLE referendum_ballot_responses (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                          referendum_id BIGINT,
+                                          ballot_response_id BIGINT,
                                           sort_order INTEGER);
