+++ resources/migrations/20150605-20-create-ballot-candidates.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE ballot_candidates (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                ballot_id INTEGER,
-                                candidate_id INTEGER);
+CREATE TABLE ballot_candidates (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                ballot_id BIGINT,
+                                candidate_id BIGINT);
