+++ resources/migrations/20150605-22-create-precinct-split-polling-locations.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE precinct_split_polling_locations (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                               precinct_split_id INTEGER,
-                                               polling_location_id INTEGER);
+CREATE TABLE precinct_split_polling_locations (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                               precinct_split_id BIGINT,
+                                               polling_location_id BIGINT);
