+++ resources/migrations/20150605-24-create-precinct-early-vote-sites.up.sql
@@ -1,3 +1,3 @@
-CREATE TABLE precinct_early_vote_sites (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                        precinct_id INTEGER,
-                                        early_vote_site_id INTEGER);
+CREATE TABLE precinct_early_vote_sites (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                        precinct_id BIGINT,
+                                        early_vote_site_id BIGINT);
