+++ resources/migrations/20150605-26-custom-ballot-ballot-responses.up.sql
@@ -1,4 +1,4 @@
-CREATE TABLE custom_ballot_ballot_responses (results_id INTEGER REFERENCES results (id) NOT NULL,
-                                             custom_ballot_id INTEGER,
-                                             ballot_response_id INTEGER,
+CREATE TABLE custom_ballot_ballot_responses (results_id BIGINT REFERENCES results (id) NOT NULL,
+                                             custom_ballot_id BIGINT,
+                                             ballot_response_id BIGINT,
                                              sort_order INTEGER);
